stock tests

+++ futures/robots/remove-parts.ts
@@ -1,4 +1,7 @@
-export const removeParts = async (parts_id: string) => {
+export const removeParts = async (parts_id: string, card_id: string) => {
+
+    if (!card_id) throw new Error(`No card_id provided`)
+
     const res = await fetch(`/api/robots/remove-parts`, {
         method: 'DELETE',
         headers: {
@@ -10,5 +13,15 @@ export const removeParts = async (parts_id: string) => {
     })
 
     if (!res.ok) throw new Error(`Could not remove parts with id: ${parts_id}`)
+
+    await fetch(`/api/user/update-user-score`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+            card_id,
+            value: -0.05,
+        })
+    })
+
     return res.json();
 };
